+++ crates/canic-core/src/env/ck.rs
@@ -1,50 +1,5 @@
 //! List of Chain Fusion (ck) canisters.
 //! Source: https://dashboard.internetcomputer.org/chain-fusion
+// The data is included so build.rs can validate the same literals.
 
-static_canisters! {
-
-    //
-    // CHAINS
-    //
-
-    BTC_MAINNET_CANISTER = "ghsi2-tqaaa-aaaan-aaaca-cai";
-    EVM_RPC_CANISTER = "7hfb6-caaaa-aaaar-qadga-cai";
-    SOL_RPC_CANISTER = "tghme-zyaaa-aaaar-qarca-cai";
-
-    // ckBTC
-    CKBTC_LEDGER_CANISTER = "mxzaz-hqaaa-aaaar-qaada-cai";
-    CKBTC_MINTER_CANISTER = "mqygn-kiaaa-aaaar-qaadq-cai";
-    CKBTC_INDEX_CANISTER = "n5wcd-faaaa-aaaar-qaaea-cai";
-    CKBTC_CHECKER_CANISTER = "oltsj-fqaaa-aaaar-qal5q-cai";
-
-    // ckETH
-    CKETH_LEDGER_CANISTER = "ss2fx-dyaaa-aaaar-qacoq-cai";
-    CKETH_MINTER_CANISTER = "sv3dd-oaaaa-aaaar-qacoa-cai";
-    CKETH_INDEX_CANISTER = "s3zol-vqaaa-aaaar-qacpa-cai";
-
-    //
-    // ERC-20
-    //
-
-    CKERC20_ORCHESTRATOR_CANISTER = "vxkom-oyaaa-aaaar-qafda-cai";
-
-    // ckUSDC
-    CKUSDC_LEDGER_CANISTER = "cngnf-vqaaa-aaaar-qag4q-cai";
-    CKUSDC_INDEX_CANISTER = "cefgz-dyaaa-aaaar-qag5a-cai";
-
-    // ckUSDT
-    CKUSDT_LEDGER_CANISTER = "xevnm-gaaaa-aaaar-qafnq-cai";
-    CKUSDT_INDEX_CANISTER = "xrs4b-hiaaa-aaaar-qafoa-cai";
-
-    // ckLINK
-    CKLINK_LEDGER_CANISTER = "g4tto-rqaaa-aaaar-qageq-cai";
-    CKLINK_INDEX_CANISTER = "gvqys-hyaaa-aaaar-qagfa-cai";
-
-    // ckOCT
-    CKOCT_LEDGER_CANISTER = "ebo5g-cyaaa-aaaar-qagla-cai";
-    CKOCT_INDEX_CANISTER = "egp3s-paaaa-aaaar-qaglq-cai";
-
-    // ckPEPE
-    CKPEPE_LEDGER_CANISTER = "etik7-oiaaa-aaaar-qagia-cai";
-    CKPEPE_INDEX_CANISTER = "eujml-dqaaa-aaaar-qagiq-cai";
-}
+include!("ck.inc.rs");