+++ crates/canic-core/src/ops/ic/mgmt.rs
@@ -111,7 +111,7 @@ pub async fn raw_rand() -> Result<[u8; 32], Error> {
 //
 
 /// Queries the NNS registry for the subnet that this canister belongs to and records ICC metrics.
-pub async fn get_current_subnet_pid() -> Result<Option<Principal>, Error> {
+pub async fn try_get_current_subnet_pid() -> Result<Option<Principal>, Error> {
     let request = GetSubnetForCanisterRequest::new(crate::cdk::api::canister_self());
 
     let subnet_id_opt = Call::unbounded_wait(*NNS_REGISTRY_CANISTER, "get_subnet_for_canister")
@@ -122,9 +122,17 @@ pub async fn get_current_subnet_pid() -> Result<Option<Principal>, Error> {
         .subnet_id;
 
     if let Some(subnet_id) = subnet_id_opt {
-        log!(Topic::Topology, Info, "get_current_subnet_pid: {subnet_id}");
+        log!(
+            Topic::Topology,
+            Info,
+            "try_get_current_subnet_pid: {subnet_id}"
+        );
     } else {
-        log!(Topic::Topology, Warn, "get_current_subnet_pid: not found");
+        log!(
+            Topic::Topology,
+            Warn,
+            "try_get_current_subnet_pid: not found"
+        );
     }
 
     Ok(subnet_id_opt)